test for Makefile

+++ libmini/qtviewer/build.sh
@@ -6,8 +6,10 @@
 # for build options type "./build.sh help"
 
 if ($1 == "clean") then
-   make distclean
-   rm -rf Makefile* qtviewer qtviewer_debug qtviewer*.app mocs* objs* rccs* uics*
+   if (-e Makefile) then
+      make distclean
+      rm -rf Makefile* qtviewer qtviewer_debug qtviewer*.app mocs* objs* rccs* uics*
+   endif
    exit
 endif
